+++ lib/rubocop/cop/mixin/multiline_literal_brace_layout.rb
@@ -8,7 +8,7 @@ module RuboCop
     module MultilineLiteralBraceLayout
       def check_brace_layout(node)
         return unless node.loc.begin # Ignore implicit literals.
-        return if node.children.empty? # Ignore empty literals.
+        return if children(node).empty? # Ignore empty literals.
 
         if opening_brace_on_same_line?(node)
           return if closing_brace_on_same_line?(node)
@@ -29,7 +29,7 @@ module RuboCop
         else
           range = Parser::Source::Range.new(
             node.source_range.source_buffer,
-            node.children.last.source_range.end_pos,
+            children(node).last.source_range.end_pos,
             node.loc.end.begin_pos)
 
           ->(corrector) { corrector.remove(range) }
@@ -38,16 +38,20 @@ module RuboCop
 
       private
 
+      def children(node)
+        node.children
+      end
+
       # This method depends on the fact that we have guarded
       # against implicit and empty literals.
       def opening_brace_on_same_line?(node)
-        node.loc.begin.line == node.children.first.loc.first_line
+        node.loc.begin.line == children(node).first.loc.first_line
       end
 
       # This method depends on the fact that we have guarded
       # against implicit and empty literals.
       def closing_brace_on_same_line?(node)
-        node.loc.end.line == node.children.last.loc.last_line
+        node.loc.end.line == children(node).last.loc.last_line
       end
     end
   end